Rename package to `runtime-introspect`
