+++ .github/workflows/cd.yml
@@ -32,12 +32,17 @@ jobs:
     name: Test source distribution
     needs: [build]
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version:
+          - '3.10'
+          - '3.14'
     steps:
     - name: Checkout Source
       uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
     - uses: astral-sh/setup-uv@e92bafb6253dcd438e0484186d7669ea7a8ca1cc # v6.4.3
       with:
-        python-version: '3.10'
+        python-version: ${{ matrix.python-version }}
         enable-cache: true
         prune-cache: false
         cache-suffix: test-sdist
@@ -49,7 +54,7 @@ jobs:
     - name: test source distribution in isolation
       run: |
         tar xzf dist/*.tar.gz
-        cd runtime-features-introspection-*
+        cd runtime_introspect-*
         uv pip install .
         uv run --no-project pytest --color=yes -ra
 
@@ -61,7 +66,7 @@ jobs:
     runs-on: ubuntu-latest
     environment:
       name: pypi
-      url: https://pypi.org/p/runtime-features-introspection
+      url: https://pypi.org/p/runtime-introspect
     permissions:
       id-token: write
     steps:
