+++ .github/workflows/ci.yml
@@ -72,7 +72,7 @@ jobs:
       if: ${{ startswith( matrix.os, 'ubuntu' ) }}
       uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
       with:
-        name: runtime-features-introspection_coverage_data-${{ matrix.os }}-${{ matrix.python-version }}-${{ matrix.prelude }}
+        name: runtime-introspect_coverage_data-${{ matrix.os }}-${{ matrix.python-version }}-${{ matrix.prelude }}
         path: .coverage.*
         if-no-files-found: ignore
         include-hidden-files: true
@@ -95,7 +95,7 @@ jobs:
 
     - uses: actions/download-artifact@634f93cb2916e3fdff6788551b99b062d0335ce0 # v5.0.0
       with:
-        pattern: runtime-features-introspection_coverage_data-*
+        pattern: runtime-introspect_coverage_data-*
         merge-multiple: true
 
     - name: Check coverage
@@ -107,7 +107,7 @@ jobs:
     - name: Upload HTML report if check failed.
       uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
       with:
-        name: runtime-features-introspection_coverage_report
+        name: runtime-introspect_coverage_report
         path: htmlcov
       if: ${{ failure() }}
 
@@ -140,4 +140,4 @@ jobs:
       run: ${{ matrix.typechecker }} src
     - name: Verify Types (pyright only)
       if: ${{ startsWith( matrix.typechecker , 'pyright' ) }}
-      run: ${{ matrix.typechecker }} --verifytypes runtime_features_introspection
+      run: ${{ matrix.typechecker }} --verifytypes runtime_introspect
