+++ README.md
@@ -1,2 +1,2 @@
-# runtime-features-introspection
+# runtime-introspect
 A small Python library to seamlessly introspect interpreter features at runtime
